refactor(orchestrator): move executeRemCycle + readiness gate to DreamService (#12069)
Operator-direct architectural correction. PR #12092 (already merged) and the
prior PR #12096 commit both added large methods directly to Orchestrator.mjs,
contradicting the recent service-decomposition refactor that thinned the
orchestrator into a coordinator + delegated logic to dedicated service files
(MaintenanceBackpressureService, CadenceEngine, DreamService, etc.).

DreamService is the natural home for executeRemCycle: it already owns the REM
pipeline (processUndigestedSessions, findUndigestedSessions, etc.), so the
typed-outcome unified entrypoint composes existing methods on `this`. The
checkProviderReadiness helper moves with it -- both depend on the runSandman
helpers (waitForProvider, createProviderFailureDiagnostic,
getGraphProviderReadinessTarget) which DreamService now imports directly.

Net effect on Orchestrator.mjs:
- DELETE async executeRemCycle({...}) and its JSDoc (~110 LOC body)
- DELETE async runProviderReadinessGate() and its JSDoc (~50 LOC body)
- DELETE imports: GraphService, createProviderFailureDiagnostic,
  getGraphProviderReadinessTarget, waitForProvider
- Refactor the periodic dream callback: stop calling local methods, delegate
  to this.dreamService.executeRemCycle({...}) and map the typed outcome to
  recordTaskOutcome
- Drive-by: fix the pre-existing `reason.reason` bug in the dream callback
  (CadenceEngine already unwraps the `{reason: '...'}` object via line 46 of
  CadenceEngine.mjs, so the callback receives the string directly -- matches
  the summary/kbSync pattern)

Net effect on DreamService.mjs:
- ADD async executeRemCycle({reason, mode, includeGoldenPath, includeDecay,
  dryRun}) returning the typed outcome envelope (completed | skipped | failed)
- ADD async checkProviderReadiness() returning {ready, diagnostic?}
- ADD imports: AiConfig + the three runSandman readiness helpers

Tests:
- DELETE Orchestrator.executeRemCycle.spec.mjs (orchestrator scope)
- ADD services/DreamService.executeRemCycle.spec.mjs with the same 9 typed-
  outcome cases, now stubbing the singleton via the beforeEach/afterEach
  swap pattern (DreamService is a Neo.setupClass singleton instance)
- Update Orchestrator.spec.mjs consumer-path case to stub
  dreamService.executeRemCycle directly + assert the orchestrator passes
  the right options envelope + maps the typed `failed` outcome to
  recordTaskOutcome with failurePhase: 'provider-readiness' + diagnostic
