+++ ai/daemons/orchestrator/services/DreamService.mjs
@@ -21,6 +21,12 @@ import MemorySessionIngestor from '../../../services/ingestion/MemorySessionInge
 import SemanticGraphExtractor from '../../../services/graph/SemanticGraphExtractor.mjs';
 import TopologyInferenceEngine from '../../../services/graph/TopologyInferenceEngine.mjs';
 import GoldenPathSynthesizer from '../../../services/graph/GoldenPathSynthesizer.mjs';
+import AiConfig from '../../../config.mjs';
+import {
+    createProviderFailureDiagnostic,
+    getGraphProviderReadinessTarget,
+    waitForProvider
+} from '../../../scripts/runners/runSandman.mjs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -255,6 +261,165 @@ class DreamService extends Base {
         } finally {
             this.isProcessing = false;
         }
+    }
+
+    /**
+     * Unified canonical REM (Sandman) cycle entrypoint. Returns a typed cycle outcome
+     * envelope so callers + observability surfaces can distinguish real work from
+     * silent no-ops and structured failures — the keystone fix for the orchestrator
+     * periodic dream path that previously mapped every non-throwing return to
+     * `completed` regardless of whether sessions were actually processed.
+     *
+     * Outcome status semantics:
+     *
+     * - `completed` — provider was ready, undigested sessions existed, processing finished
+     *   without throw. `sessionsProcessed` carries the pre-call count.
+     * - `skipped`   — ran successfully but did no work (concurrent-invocation guard,
+     *   zero undigested sessions, dry-run, etc). `skipReason` carries the diagnostic.
+     * - `failed`    — provider readiness gate rejected OR in-pipeline throw caught.
+     *   Either `diagnostic` (for provider failure) or `error` (for throws) is populated.
+     *
+     * Lease ownership stays with the caller. The orchestrator periodic dream path runs
+     * inside `MaintenanceBackpressureService`'s lease; the standalone Sandman CLI runner
+     * acquires its own lease via `withHeavyMaintenanceLease`. Re-acquiring inside this
+     * method would double-lease and deadlock.
+     *
+     * @param {Object}  [options]
+     * @param {String}  [options.reason]            Coordination string for logging + state model (e.g. `periodic-dream:3600000`).
+     * @param {String}  [options.mode='periodic']   `'periodic' | 'manual' | 'cli'`.
+     * @param {Boolean} [options.includeGoldenPath=false] Reserved for the Sub 5 refresh helper; periodic path leaves it false.
+     * @param {Boolean} [options.includeDecay=true] When true, runs `GraphService.decayGlobalTopology()` as the cycle-finalization step (24-hour Algorithmic Lock self-skips when not due).
+     * @param {Boolean} [options.dryRun=false]      Probe-only mode; short-circuits to `skipped` after the readiness gate passes.
+     * @returns {Promise<Object>} typed outcome envelope (see status semantics above).
+     */
+    async executeRemCycle({
+        reason,
+        mode             = 'periodic',
+        includeGoldenPath = false,
+        includeDecay     = true,
+        dryRun           = false
+    } = {}) {
+        const startedAt = new Date();
+        const runId     = `rem-${startedAt.toISOString()}-${Math.random().toString(36).slice(2, 8)}`;
+
+        const baseOutcome = {
+            runId,
+            reason,
+            mode,
+            startedAt        : startedAt.toISOString(),
+            completedAt      : null,
+            durationMs       : null,
+            sessionsProcessed: null,
+            diagnostic       : null,
+            skipReason       : null,
+            error            : null
+        };
+
+        const finalize = (status, extras = {}) => ({
+            ...baseOutcome,
+            ...extras,
+            status,
+            completedAt: new Date().toISOString(),
+            durationMs : Date.now() - startedAt.getTime()
+        });
+
+        const gate = await this.checkProviderReadiness();
+        if (!gate.ready) {
+            return finalize('failed', {diagnostic: gate.diagnostic});
+        }
+
+        if (dryRun) {
+            return finalize('skipped', {skipReason: 'dry-run requested'});
+        }
+
+        if (this.isProcessing) {
+            return finalize('skipped', {skipReason: 'dreamService.isProcessing already true (concurrent invocation)'});
+        }
+
+        let sessionCount = 0;
+        try {
+            const undigested = await this.findUndigestedSessions();
+            sessionCount = Array.isArray(undigested) ? undigested.length : 0;
+        } catch (e) {
+            return finalize('failed', {
+                error: {message: `findUndigestedSessions threw: ${e?.message || e}`, stack: e?.stack}
+            });
+        }
+
+        if (sessionCount === 0) {
+            if (includeDecay) {
+                try {
+                    await GraphService.decayGlobalTopology();
+                } catch (e) {
+                    return finalize('failed', {
+                        error            : {message: `decayGlobalTopology threw on zero-session path: ${e?.message || e}`, stack: e?.stack},
+                        sessionsProcessed: 0
+                    });
+                }
+            }
+            return finalize('skipped', {sessionsProcessed: 0, skipReason: 'no undigested sessions'});
+        }
+
+        try {
+            await this.processUndigestedSessions();
+
+            if (includeDecay) {
+                await GraphService.decayGlobalTopology();
+            }
+
+            if (includeGoldenPath) {
+                logger.info('[DreamService] executeRemCycle: includeGoldenPath flag set but the dedicated refresh helper is deferred to its own ticket; periodic dream path does NOT refresh goldenPath in this cycle');
+            }
+
+            return finalize('completed', {sessionsProcessed: sessionCount});
+        } catch (e) {
+            return finalize('failed', {
+                sessionsProcessed: sessionCount,
+                error            : {message: String(e?.message || e), stack: e?.stack}
+            });
+        }
+    }
+
+    /**
+     * Probes the configured graph provider before invoking a graph-heavy REM cycle.
+     * Returns `{ready: true}` when the configured provider answers the HTTP probe,
+     * `{ready: false, diagnostic}` when the provider is unsupported or the readiness
+     * loop exhausts retries.
+     *
+     * Probe parameters flow from `aiConfig.orchestrator.providerReadiness` verbatim
+     * (no module-level fallbacks). Daemon-context invocations suppress the dot-progress
+     * writer used by the CLI runner.
+     *
+     * @returns {Promise<{ready: true} | {ready: false, diagnostic: Object}>}
+     */
+    async checkProviderReadiness() {
+        const readinessConfig = AiConfig.orchestrator.providerReadiness;
+        const target          = getGraphProviderReadinessTarget();
+
+        if (!target.supported) {
+            return {
+                ready     : false,
+                diagnostic: createProviderFailureDiagnostic({
+                    reason: 'UNSUPPORTED_GRAPH_PROVIDER'
+                })
+            };
+        }
+
+        const waitResult = await waitForProvider({
+            attempts : readinessConfig.attempts,
+            delayMs  : readinessConfig.delayMs,
+            timeoutMs: readinessConfig.timeoutMs,
+            output   : {write: () => {}}
+        });
+
+        if (!waitResult.running) {
+            return {
+                ready     : false,
+                diagnostic: createProviderFailureDiagnostic({waitResult})
+            };
+        }
+
+        return {ready: true};
     }
 
     /**
